pyproject.toml
[project]
name = "hertopia-musica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colormath>=3.0.0",
    "evdev>=1.9.3",
    "mido>=1.3.3",
    "numpy>=2.4.2",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "rich>=14.3.2",
    "tk>=0.1.0",
]

[tool.basedpyright]
typeCheckingMode = "standard"
stubPath = "typings"
